requirements.txt
asgiref==3.2.10
Django==3.1.2
django-bootstrap-modal-forms==2.0.0
django-cleanup==5.1.0
django-countries==6.1.3
django-extensions==3.1.0
django-money==1.1
django-richtextfield==1.6
django-widget-tweaks==1.4.8
graphviz==0.15
Pillow==7.2.0
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
py-moneyed==0.8.0
pydotplus==2.0.2
pyparsing==2.4.7
pytz==2020.1
sqlparse==0.4.0
